+++ src/extentions/Trigger2D.cpp
@@ -1,7 +1,7 @@
 #include "Trigger2D.h"
 #include <Arduino.h>
 
-Trigger2D::Trigger2D(int x1, int y1, int x2, int y2)
+Trigger2D::Trigger2D(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2)
 {
     setSize(x1,  y1,  x2,  y2);
     inside = false;
@@ -17,7 +17,7 @@ void Trigger2D::update(Transform &transform)
     update(transform.getPositionX(), transform.getPositionY());
 }
 
-void Trigger2D::update(int x, int y)
+void Trigger2D::update(uint16_t x, uint16_t y)
 {
     bool currentlyInside = isInside(x, y);
 
@@ -37,7 +37,7 @@ void Trigger2D::update(int x, int y)
     }
 }
 
-bool Trigger2D::isInside(int x, int y)
+bool Trigger2D::isInside(uint16_t x, uint16_t y)
 {
     return (x >= minX && x <= maxX && y >= minY && y <= maxY);
 }
@@ -68,7 +68,7 @@ void Trigger2D::OnTriggerStay()
 }
 
 
-void Trigger2D::setSize(int x1, int y1, int x2, int y2)
+void Trigger2D::setSize(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2)
 {
     minX = min(x1, x2);
     minY = min(y1, y2);
