+++ src/extentions/Trigger2D.h
@@ -5,10 +5,10 @@
 class Trigger2D
 {
 private:
-    int minX;
-    int minY;
-    int maxX;
-    int maxY;
+    uint16_t minX;
+    uint16_t minY;
+    uint16_t maxX;
+    uint16_t maxY;
     bool inside;
 
 protected:
@@ -18,17 +18,17 @@ protected:
 
 public:
     bool isActive = true;
-    Trigger2D(int x1, int y1, int x2, int y2);
+    Trigger2D(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2);
 
     void update(Transform &transform);
     void update(Vector2Byte &vector2Byte);
-    void update(int x, int y);
+    void update(uint16_t x, uint16_t y);
 
     bool isInside(Transform &transform);
     bool isInside(Vector2Byte &vector2Byte);
-    bool isInside(int x, int y);
+    bool isInside(uint16_t x, uint16_t y);
     
 
-    void setSize(int x1, int y1, int x2, int y2);
+    void setSize(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2);
 
 };
